server/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
django-cors-headers==4.8.0
